+++ lib/cocoapods/installer.rb
@@ -331,7 +331,7 @@ module Pod
         all_projects_by_pod_targets.merge!(pods_project_by_targets) if pods_project_by_targets
         all_projects_by_pod_targets.merge!(projects_by_pod_targets) if projects_by_pod_targets
         all_projects_by_pod_targets.each do |project, pod_targets|
-          generator.configure_schemes(project, pod_targets)
+          generator.configure_schemes(project, pod_targets, pod_project_generation_result)
         end
       end
     end